packages/mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-server"
version = "0.1.0"
description = "FastMCP server for Tiger Brokers API integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.10.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.28.1",
    "backoff>=2.2.0",
    "shared",
    # Tiger SDK dependencies (installed separately from local SDK)
    "simplejson>=3.19.0",
    "delorean>=1.0.0",
    "pandas>=2.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "pyasn1>=0.4.8",
    "rsa>=4.9.0",
    "stomp.py>=8.0.0",
    "getmac>=0.9.0",
    "cryptography>=41.0.0",
    "jproperties>=2.1.0",
    "protobuf>=4.24.0",
    "google-cloud>=0.34.0",
    "urllib3>=2.0.0",
]

[project.scripts]
tiger-mcp-server = "mcp_server.cli:main"
tiger-mcp-stdio = "mcp_server.main:main"
tiger-mcp-sse = "mcp_server.main:main_sse"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
]

[tool.uv.sources]
shared = { workspace = true }

[dependency-groups]
dev = [
    "pytest-mock>=3.14.1",
    "pytest-cov>=6.2.1",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.28.1",
]
